requirements.txt
# FastAPI et serveur
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Base de données
sqlalchemy>=2.0.23
pymysql>=1.1.0
cryptography>=41.0.7
alembic>=1.12.1

# Authentification et sécurité
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-dotenv>=1.0.0

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Tests
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2

# CORS
fastapi-cors>=0.0.6